# docs/swagger.yaml
basePath: /api/
definitions:
  entity.Authenticate:
    properties:
      email:
        example: james@yahoo.com
        type: string
      password:
        example: password!@#123
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
host: localhost:8000
info:
  contact:
    email: arshamdev2001@gmail.com
  termsOfService: http://swagger.io/terms/
  title: UrlShortener Swagger Document
  version: "1.0"
paths:
  /auth/login/:
    post:
      consumes:
      - application/json
      description: Login user with email and password
      parameters:
      - description: Login Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.Authenticate'
      responses: {}
      summary: Login User
      tags:
      - Auth
  /auth/signup/:
    post:
      consumes:
      - application/json
      description: register user with email and password and sending verfication email
      parameters:
      - description: Register Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.Authenticate'
      responses: {}
      summary: Register User
      tags:
      - Auth
  /auth/verify-user/:
    post:
      consumes:
      - application/json
      description: send verification link key in query param to active user
      parameters:
      - description: verification query param
        in: query
        name: key
        required: true
        type: string
      responses: {}
      summary: Verify User Account
      tags:
      - Auth
  /user/change-password/:
    post:
      consumes:
      - application/json
      description: Change user password with current and new password
      responses: {}
      summary: Change Password
      tags:
      - User
  /user/delete-user/:
    delete:
      consumes:
      - application/json
      description: Delete user record from db
      responses: {}
      summary: Delete User
      tags:
      - User
  /user/me/:
    get:
      consumes:
      - application/json
      description: Get user data information
      responses: {}
      summary: Get User
      tags:
      - User
schemes:
- http
- https
securityDefinitions:
  Authorization:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"